Perfect Hit
Ring Particle for perfect hit
- Attach to sword
- Particle faces players view

Simple idea; when player lands a perfect hit (to the beat) a simple ring effect can spawn on the player sword.

Other idea; Ring effect can be adjusted to spwan multiple rings (i.e. 2 rings instead of the one).
Copy and paste the ring particle, adjust so that it spawns 1 ring, then 2 rings, then 3 rings and set up a script so that on the first perfect hit 1 ring is spawned, second perfect hit 2rings are spawned, etc.